sync-gerrit-review.py refactoring
The idea is to refactor sync-gerrit-review.py to get something easier to
handle. I wish to be able to define some stream of components like: -
the poodle snapshot from a Yum repository - the last git commit from a
Github repository - the git review from a Gerrit project - etc The test
is still hard-coded and called 'khaleesi-tempest'. I think we should be
able to overwrite some settings directly in the jobdefinitions, to do
so, we will need a jobdefinitions.data key.

Change-Id: I7238f1b30f01c2cb967fdf60eebcf2b5641ac224