Build packages for various platforms (Fedora, EL)
Currently packages are built only for the platform this script is run
on. Meaning that if the node this script is run on is a Fedora-23 only
packages for Fedora-23 will be created, same apply for Fedora-22 or
CentOS-7.

This commit aims for this script to build the packages for a given set
of platform we want to handle. The platform set is :

  * fedora-22-x86_64
  * fedora-23-x86_64
  * epel-7-x86_64 (CentOS 7)

Change-Id: I6946362c1f6690b984361707ded69db65bfa3e74