Design an inheritance base component arch
Currently all the method for componenttype are store in DCIClient.
This doesn't scale right as we will add method for the other components;
such as user, component, jobdefinitions, etc...

This review aims to set an orgnanization in order to have a clean tree
for the upcoming components we will create. Each component being its own
class both at the Click level and at the handler level.

Change-Id: I06152e53ba5336c7fa38cc126c37dc29305de50a