handlers: replace show by get, add embed and where attributes
The subclasses were calling the show method when really what they wanted
to call is the super get method. I think that it worked due to the
method signature but this wasn't correct.

This commit allows the get method to pass extra embed and where
parameters to query the API.

Change-Id: I18d87b231c0e2fcfe6fd64020a37ad09f94e636e