consumer: Implement the notion of agent and feeder
This commits brings new components in the python-dciclient. It brings
the notion of agents and feeders.

Feeders : Script in charge of detecting (either by an event or by
polling a system) something new to test, and feed the DCI server with
those new components

Agents : Script that are run on the customer platform, to perform a
specific action against a components feeded by the feeder

Change-Id: Ic5306446e7f0e865c7e44e78b700f15f198b8247