Update patch set 7
Patch Set 7: Code-Review-1

This is not a fix.

If the item key is missing it does not means that it's an error, take a look at delete() for instance.

The fix is to force the print of the json result in case of an http error code.

I don't understand the tests.

Patch-set: 7
Reviewer: Gerrit User 29 <29@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=-1