DCI log handler for PyLogger
This package has been created by @ylamgarchal for the
python-rdo-m-helper package. As discussed, it deserves to be merged in
this package instead.

This module has been slightly adjusted to allow to bind log() with the dynamic
create of new jobstate entry.

The unit-tests make use of the where parameter of list() that wasn't
implemented yet in the lib.

Co-Authored-By: Yassine Lamgarchal <yassine.lamgarchal@redhat.com>

Change-Id: I5c71f420031032ad85cfd8b328765d1ba2309f4a