packaging: Sign our packaged during build process
This commit aims to sign our packaged during the build process.
We rely on the sign[1] plugin offered by mock to run the rpm --addsign
command. In order to sign them we need to specify two custom rpm macros
that are :

  * %_signature
  * %_gpg_name

Finally we wrap everything around expect so it is non-blocking

[1]
https://fedoraproject.org/wiki/Mock/Plugin/Sign?rd=Subprojects/Mock/Plugin/Sign

Change-Id: I57b2ac8e01ec8f8818d4117831ce8c02f486d40e