jobdefinitions and tests: Create a many to many relationship
Currently a jobdefinition can be attached only a single job, creating a
defacto 1 to 1 relationship.

However for a given jobdefinition an agent might want to be able to run
a series of test and not just one.

Taking the OpenStack example, one might run : Tempest, Rally, and an
in-house test suite.

This commits modify this relation between jobdefinition and test making
it many to many. Meaning a jobdefinition can have many tests, and
reciprocally the same tests can be attached to various job definitions.

Change-Id: I8dcaedf94e41548f8d7296190a1dbfefff238b83