Update patch set 2
Patch Set 2:

By leaving this task to the agent and not the control-server, we do loose the ability to enforce this number of retries for a specific jobdefinition.

Let say, we want OpenStack's job to run 5 time successfully to be considered 'Success', as we control the control-server we can enforce this value. If it is done from the agent side we just loose this, and if an agent operator decide to put 1 it will work for her

Patch-set: 2
Reviewer: Gerrit User 115 <115@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=0