Update patch set 1
Patch Set 1: Code-Review-1

Disagree with how the thing are done here.

dciclient is the generic library it should not 'import' clases that come from a speficic 'agent' here tripleohelper.

This feature belongs to the agent that will specifically run tripleohelper, but not dciclient as it purpose is to be agent-agnostic.

Patch-set: 1
Label: Code-Review=-1