Update patch set 1
Patch Set 1: Code-Review-1

If I am not mistaken we can safely remove both Conflicts and Obsoletes instruction.

This change has an impact during installation process. Now that dci-feeders lives in its own repo, that we provide the packages in our own repo and that we never publicly released the RPM for it, the incompatible combination should never happen.

If it does, simply removing and reinstalling python-dciclient/dci-feeders will be enough, but removing those two lines from the spec file will maintain it clearer.

Patch-set: 1
Label: Code-Review=-1