Update patch set 1
Patch Set 1:

(1 comment)

Small bug in the code + test missing. Once bug fixed it works as expected on local workstation.

Should we add an options to limit the last jobstates we want to retrieve? Create a --ouput option to save output to a file (I know > works, but might be more user friendly the --output)

Patch-set: 1