change export_control default to True
Change the default value of the export_control field fron False to True.
This means a newly create component will know be dispatched to agent
when they request a new job.
The previous behavior was to enforce the use of --export-control
parameter. But in the vast majority of our use cases, we don't need the
ignore components.

Change-Id: Ifadc3201c7c5ceec8f47d280b6bcd594fe522df3