file-list now return all the files of a job
Until I317043fb5292d6b09e8f60d060cd36cc4e7db26f, some files were only
associated to a jobdefinition, so were also associated to the job.
This, depending on the way the file was created.

Now, all these files are still associated to the jobdefinition and
the job.

Change-Id: I792cb188e6ea65b9d7b79a84ebb9e25a793effb7