tox: Allow to disable pifpaf
Since daeb5f6598e198d05e9d879afa4382dcbc41e497 we are using pifpaf to
manage the postgresql database instead of shell script.
But we lose the possibility to use an external db (like we do with
containers).
This patch allows to disable the startup of a local postgresql server
used by pifpaf.

Change-Id: Id487b9103a05de70abba708b82e1f53336ad5153