Update patch set 1
Patch Set 1: Code-Review-1

OSP comes with RHEL python-setuptools (0.9.8) where RDO package python2-setuptools (22.0.5).

The problem is that we build dci-agent el7 on CentOS/RDO. During the build the dci-agent script is generated with setuptools 22.

When we try to run dci-agent on RHEL, it cannot reuse the same setuptools version and we get this nice backtrace. A simple solution is to upgrade setuptool to 22 but it's a workaround.

IMO a better solution IMO is to use a RHOS puddle when we build our packages.

Patch-set: 1