test: Disable test_component_status
This is a temporary fix to get review 5771 working. Because we are
testing dciclient tox in api review, we can't get the review
in success only by updating the test in the client.
We need first to disable it, merged the api review and then reenable
it with the right value.

Change-Id: I2b38ebcd898a729fa4cd4c5b57d7c1e2b34e5f1b