test: Reenable test_component_status
This is a temporary fix to get review 5771 working. Because we are
testing dciclient tox in api review, we can't get the review
in success only by updating the test in the client.
We need first to disable it, merged the api review and then reenable
it with the right value.

Change-Id: Id58658dc7c6f0237aa4cf7f3698367bfb58f93de
Depends-On: I20fa6471d181efc2626a15347dadbdc48b7ae99e