Enable two new runners: runner_user and runner_user_admin
Currently all the test are run with the same context: admin.
So everytime we call runner.invoke() we call the CLI as the admin user.

This prevents us from doing authorization tests, as admin can do
everything. This reviews creates two new runners:

  * runner_user
  * runner_user_admin

Thos two runners make the queries in their proper context runner_user as
a regular user and runner_user_admin as an admin of his team.

Change-Id: Ifacacef1f5caf393f1f0f6aa56cafb2a14c93438