Implement the --active/--no-active on ressources
Since the move to the state field on the Control-Server side more
resources now support this field.

The field state can only be set to three values:

  * active
  * inactive
  * archived

The archived value for this field is achieved by deleting the resource,
so one should never manually set a the state field to this value. Hence
leaving us with the user being able to state if a resource is either
active or inactive.

This boolean flag has been spreaded to the resource that has the state
field but was lacking support in the dcictl bin.

Depends-On: I416dcc4f579965382e2cc11d481b1d47d0a28d27
Change-Id: I048b93fa00c2fa17b9af56749b02a36ae69a8b5a