Update patch set 3
Patch Set 3: Code-Review-1

If I understood correctly the purpose of this script is to prepare jobs deletion.
Still according to the map, everyone has to agree on what a bad job is.
And for me there should be a test for scripts used for destructive actions.

-1 just to be sure we don't merge this without a discussion
code is ok

Patch-set: 3