Update: Fix boolean vs is_flag in parameters
Currently, parameters that during creation are either true or false (ie.
active, notification, export-control) are set as boolean even during
update.

This leads to a value to be set *all the time*, even when no specific
value has been specified.

If we have an 'inactive' user, for example and decide to update his
name, when we run 'dcictl user-update --name foo' the flag
--active/-no-active will be set as True (--active) as it is either True
or False, can not be anything else, leading to the user to move from
inactive to active while never asked.

Adding 'default=None' to these flags when updating fixes this behaviour.

Change-Id: If52d3b8135192656b2c08861a142c4820706d9e6
Depends-On: I9e2fbd0be0b64b8ed2e34ead4aa54f04938e4c58