Update patch set 5
Patch Set 5: Code-Review-1

I guess this sample was not written out of pure boredom and there's a will to use it.

I'd be reluctant to deleting such jobs as this still carries information. For example. Jobs stuck for more than 10 hours have a meaning.
For some remotecis it could mean that the platform is busy doing something else if it is shared for example. For some others it could mean that a network issue is slowing things down.

For some other cases, it could be a sign of jobs blocked for any other unknown reason (BUG in our client/server ?). Deleting such jobs would result in hiding these problems and delaying a potential fix from us.

Patch-set: 5