Update patch set 1
Patch Set 1: Code-Review+1

I would rather prefer to see this function as deprecated because it's not really consistent with the rest of the API and the format tends to change silently quiet often actually. This is the kind of thing that can trigger subtle issues. I would prefer that the users just create the request that they need with the regular API.

+1 because the looks ok.

Patch-set: 1
Label: Code-Review=+1