signatureauth: remove duplicated / from the path
Depending on the configuration, dci_cs_url may have a trailing slash.
e.g: https://api.distributed-ci.io/

So far so good, this is a valid URL. The problem happens when build a
full URL from that and ask urlparse to extract the path. urlparse
won't duplicate the / and will happily return a string starting with
a //. This is not what the server expect. The client authentication
request will be refused. Poor client :(

The Werkzeug test client does not let us test an URL with a // and
return 404. So the test just ensure we never accesss a trailing /
in the server URL.

Change-Id: I42e5eb7f4c2f102d548c9c507fcfd3c3fa27ed85