Team: Allow a user to attach tests to team
Tests should be able to be attached at three different levels: RemoteCI,
Team, Topic. The ability to add tests to team was missing from the API
and the CLI.

This commit implements the feature of being able to manage tests with
regard to team. It enable the following CLI commands:

  * team-attach-test
  * team-list-tests
  * test-unattach-test

Change-Id: I321da79295c862ebbea0da2bd4ee99876d24a4e7