requests: set a timeout
By default, requests does not set a time. As a consequence of this, the
lib will hang as soon as the communication has failed. This commit set
the time to 30s.

At this point, Python's requests lib does not offert a way to set a
default timeout:
See: https://github.com/requests/requests/issues/3070

This commit also moves the download() function our of the utils module.

Change-Id: I7d58fa3bb67aa8d55ce54062c8b45968dd88207f