Update patch set 41
Patch Set 41:

(1 comment)

@Guillaume: Because you don't want to see it or something else :), in the first 22 patchsets i tried to reuse dci-control-server's scripts within dci-client-tox job.

The issue we meet is that, in order to run tests from the SPEC file, Mock needs to find these scripts within the chroot which is not possible from dci-control-server repo since they must be referenced in MANIFEST.in.

I worked on that with Cédric and Spredzy, so unless you find a better solution, this duplication is necessary to move forward :).

thanks

Patch-set: 41