get_full_data returns json object like all other
The get_full_data() method returns a non-json object while the other
calls do return a json-object. This has the consequence of having to
have two different behavior whether we call get_full_data() or another
call.

This commit aims to standardize that so we can adopt a single behavior
independantly of the method we call from the bindings.

Change-Id: I40813ad1516380cb6462662a320acd63a073d6d2