py3k: fix use of decode
Requests compat.basestring is an alias to (str, bytes) on python3.
So some methods testing if a string is an instance of compat.basestring
will try to decode the string which will fail on with unicode string.

Use instead compat.bytes which will correctly points to str on python2
and bytes on python3.

Change-Id: If506050534e078581c7bf4442bd18befe789ff9f