Refactor tags creation after server update
tags specific endpoint will be removed.
To create a tag we need to do an update to the tags field of a component.

It allows us to return the component created.
We need this update in the feeder to keep the etag updated during the process.
Today add_tag_to_resource doesnt check the etag and return None

Change-Id: Ib4c5ac1afe40e371486994c6b90d5385c26c40b4