Explicitly require python36-requests on el7
python3-requests doesn't exists on el7 so the package manager choose which version of python3X-requests it should install.
Using yum on el7 install python36-requests.
Using dnf on el7 install python34-requests.
In this patch we explicitly set the version to python36-requests on el7 to avoid this issue.

Change-Id: I3d2727b12bf95e9968b597523c77685adcc8b385