Revert "add the "query" parameter"
This reverts commit 575eba877a8c63a754d810f1412df0f8a6ef4a12.

Reason for revert: doesn't work with hmac auth

Query is empty for extra shell command like create_component.
Empty query params are removed from the HMAC auth mechanism but added in those scripts.
So scripts are not working because of an HMAC authentication error.

Revert this change. We will reapply this patch with a non regression test

Change-Id: I19c56c076446fc13fb317fb9d705a54649c3bdbc