retries: enable retries on bad status codes
In addition to connection errors, urllib3.util.Retry will retry *only* for
explicitly listed status codes

Change default max-retry count to 10 (was 0)
Set the same for the various `build_dci_context…` functions (was 80).

The feature in responses that allows to test retries only exists in
versions for python >= 3.7, run this test only for python3.9.

Change-Id: I71c686a945eafefeb9b7e5ae1a5f97bcdb66811a
Depends-On: https://softwarefactory-project.io/r/c/dci-control-server/+/32315